+++ src/main/java/org/veupathdb/service/eda/ss/model/reducer/formatter/MultiValueFormatter.java
@@ -3,7 +3,7 @@ package org.veupathdb.service.eda.ss.model.reducer.formatter;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import org.gusdb.fgputil.json.JsonUtil;
 import org.veupathdb.service.eda.ss.model.reducer.FormattedTabularRecordStreamer;
-import org.veupathdb.service.eda.ss.model.reducer.UnformattedTabularRecordStreamer;
+import org.veupathdb.service.eda.ss.model.reducer.ValueStream;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -11,7 +11,7 @@ import java.util.List;
 public class MultiValueFormatter implements TabularValueFormatter {
 
   @Override
-  public byte[] format(FormattedTabularRecordStreamer.ValueStream<byte[]> stream,
+  public byte[] format(ValueStream<byte[]> stream,
                        long idIndex) {
     List<byte[]> allRecords = new ArrayList<>();
     while (stream.hasNext() && stream.peek().getIdIndex() == idIndex) {
@@ -35,7 +35,7 @@ public class MultiValueFormatter implements TabularValueFormatter {
   }
 
   @Override
-  public String formatString(UnformattedTabularRecordStreamer.ValueStream<String> stream, long idIndex) {
+  public String formatString(ValueStream<String> stream, long idIndex) {
     List<String> allRecords = new ArrayList<>();
     while (stream.hasNext() && stream.peek().getIdIndex() == idIndex) {
       allRecords.add(stream.next().getValue());
