+++ src/main/java/org/veupathdb/service/eda/ss/model/reducer/formatter/SingleValueFormatter.java
@@ -1,19 +1,16 @@
 package org.veupathdb.service.eda.ss.model.reducer.formatter;
 
-import org.veupathdb.service.eda.ss.model.reducer.FormattedTabularRecordStreamer;
-import org.veupathdb.service.eda.ss.model.reducer.UnformattedTabularRecordStreamer;
-
-import java.util.List;
+import org.veupathdb.service.eda.ss.model.reducer.ValueStream;
 
 public class SingleValueFormatter implements TabularValueFormatter {
 
   @Override
-  public byte[] format(FormattedTabularRecordStreamer.ValueStream<byte[]> stream, long idIndex) {
+  public byte[] format(ValueStream<byte[]> stream, long idIndex) {
     return stream.next().getValue();
   }
 
   @Override
-  public String formatString(UnformattedTabularRecordStreamer.ValueStream<String> stream, long idIndex) {
+  public String formatString(ValueStream<String> stream, long idIndex) {
     return stream.next().getValue();
   }
 }
