+++ src/main/java/org/veupathdb/service/eda/ss/model/reducer/formatter/TabularValueFormatter.java
@@ -1,13 +1,11 @@
 package org.veupathdb.service.eda.ss.model.reducer.formatter;
 
 import org.veupathdb.service.eda.ss.model.reducer.FormattedTabularRecordStreamer;
-import org.veupathdb.service.eda.ss.model.reducer.UnformattedTabularRecordStreamer;
-
-import java.util.List;
+import org.veupathdb.service.eda.ss.model.reducer.ValueStream;
 
 public interface TabularValueFormatter {
 
-  byte[] format(FormattedTabularRecordStreamer.ValueStream<byte[]> stream, long idIndex);
+  byte[] format(ValueStream<byte[]> stream, long idIndex);
 
-  String formatString(UnformattedTabularRecordStreamer.ValueStream<String> stream, long idIndex);
+  String formatString(ValueStream<String> stream, long idIndex);
 }
